psplash: Change stuff based on the magic resin feature
Is important to keep our changes as scalable as possible in the sense that a
user might keep stuff in his bblayers without actually wanting resin changes.
For example, if a user forgets the layer in the bblayers and he is using
another distro on top of which is baking another bsp we shouldn't mess up with
his work. If a user really wants resin flavor, he should either use the resin
distro or inject resin DISTRO_FEATURE.

Signed-off-by: Theodor Gherzan <theodor@resin.io>