resin-init-rpi: Used to be rpi-init
The resin packagegroup needs a resin-init package that should be brought by every
machine. To keep the name consistency I renamed the recipe.
As well this recipe needs to provide resin-init and is obviously only for raspberrypi,
hence COMPATIBLE_MACHINE.
Remove the version in the recipe name as we don't have any for this package.
I can reset PR with this change.

Signed-off-by: Theodor Gherzan <theodor@resin.io>