resin-supervisor: Systemd services usage and recipes unification
This is a big commit but because it functionality is indivisible, I treat
it as one commit.

* Remove rce-run-supervisor as we handle supervisor container startup with a
systemd service - resin-supervisor.service.
* Remove resin-device-update as we handle this as a separate systemd service.
As well, given that this is a functionality for resin-supervisor, we treat this
service in that recipe. Update service: update-resin-supervisor.service.
* Define a systemd service for opening and maintaining a socket on host OS for
communication with supervisor. As well, a timer systemd service was defined
which triggers the service every 5 minutes.
* Most services have a prepare*.service. These are used as a dependency for a
specific service which need additional preparations (ex: mounts). We can't
do that as ExecStartPre as these operations take a while.
* Systemd service are installed and enabled for multi-user.target.

Signed-off-by: Andrei Gherzan <andrei@resin.io>