initramfs-live-boot: Add custom initramfs init script
For NUC support, we add our modified init script to be used when booting
from an external USB key with the resin-image-flasher rootfs.

The initramfs is necessary as we do not know at compile time what entry in /dev/
our USB key will get assigned. Thus, we boot into an initial rootfs from this initramfs,
we wait for udev to create the filesystem label entries in /dev/disk/by-label/ and after
that we can mount by label the real rootfs from the USB media.

Signed-off-by: Jon Grisky <jon@resin.io>