resin-device-register: Use oneshot service
This fixes the registration racing issue between supervisor and host OS.
Supervisor is configured to start after this service but, if registration
service is simple, systemd doesn't wait for it to finish. Oneshot type
instructs systemd to wait for the registration script to exit before proceeding
with next services (ex: supervisor).

Signed-off-by: Andrei Gherzan <andrei@resin.io>