resin-mounts-flasher: Use a volatile config file
The flashing devices needs to be left clean for flashing additional targets.
For this, we need to leave config.json intact on the flashing device.

In order to achive this we define/modify 3 services:
NEW : mnt-conforig.mount : mounts the config partition in /mnt/conforig
|
V
NEW : temp-conf.service : copies the config file from /mnt/conforig to /tmp/conf
|
V
MODIFIED : mnt-conf : mount /tmp/conf in /mnt/conf - we do this so that all the other
services will use the standard /mnt/conf path for searching the config file.

Signed-off-by: Andrei Gherzan <andrei@resin.io>