resin-supervisor-disk: Fix repository grep
Instead of doing "docker images <repository>" and then grep for what we are
interested in, we should just do a "docker images" and search through that output.

The issue we are trying to fix is that on different build Operating Systems, "docker images"
will prepend the string "docker.io" to this command's output so our current way we grep will
just return nothing.

Signed-off-by: Florin Sarbu <florin@resin.io>