meta-resin-common/conf/layer.conf: Set IMAGE_ROOTFS_SIZE as zero
We make sure this var is zero when meta-resin-common is included
in the build.
This serves as a failsafe default to specify a dummy size of the rootfs.
The actual size of the root filesystem will be adjusted automatically by
the build system, based on the actual packages included in the rootfs.

The variable value can still be changed by the user by setting it in the
build's local.conf file.

Signed-off-by: Florin Sarbu <florin@resin.io>