resin-supervisor-disk: Make sure we don't have the ':' character in the PV variable.
Docker 1.10 uses a new content-addressable storage for
images and layers. As a consequence, "docker inspect" will
return the id's prefixed with the string "sha256:".

Having the ':' character in resin-supervisor-disk's PV variable
will result in the ${B} variable also contain ":".
"docker run" will fail when trying to use bindmounts referring
to this B variable as it will end up parsing the ':' character
multiple times thus making it incorrectly interpret the new
mount point directory as a mount option instead.

This is the code we are targeting with this fix:

docker run --rm --privileged -e PARTITION_SIZE=${PARTITION_SIZE}
-e TARGET_REPOSITORY=${TARGET_REPOSITORY} -e TARGET_TAG=${SUPERVISOR_TAG}
-v /sys/fs/cgroup:/sys/fs/cgroup:ro -v ${B}:/export looper

Signed-off-by: Florin Sarbu <florin@resin.io>