connman: Change functionality so we can use dnsmasq
In order to use dnsmasq with connman, we needed to have connman not modify
/etc/resolv.conf (this will contain now only 127.0.0.1 and will be managed
by our DNS cache and forwarder - dnsmasq) so we have it write to /etc/resolv.dnsmasq
instead, which is read by dnsmasq to establish to what nameservers it will forward to.

We also patch connman to allow more than MAXNS entries in the new /etc/resolv.dnsmasq
file so that it can accomodate cases when more than MAXNS (3) entries are received by
both DHCP and connman configuration file.

Another change we did is not have connman re-order the nameserver entries in the resolv file.
It will now write to this files the entries in the order it received them (so first entries are
from the connman config file followed by entries received through DHCP in the order the DHCP
server sent them).

Signed-off-by: Florin Sarbu <florin@resin.io>