packagegroup-resin.inc: Remove RESIN_STAGING_ADDONS
There can be situations where a missed dependency in production is brought in
by a staging package hiding the issue. As well it is a good practice to try and
keep both staging and production as much as possible identical in terms of
packages installed.

This change also removes RESIN_ADDONS and expands its content in RDEPENDS
directly.

Signed-off-by: Andrei Gherzan <andrei@resin.io>