resin-image-flasher.bb: Move resin-image sdcard file to rootfs
This commit moves the resin-image sdcard file from the flash-boot partition into
the flash-root rootfs partition. This way we avoid having a big flash-boot partition,
containing files that are not relevant to the boot process.

Also, we shrink the rootfs update partition to the default alignment, as this is not used
by the flasher images.

Signed-off-by: Florin Sarbu <florin@resin.io>