resin-connman-conf: Modify PreferredTechnologies order
We change PreferredTechnologies to "wifi,ethernet" as to make sure we always
get an IP address on the wireless interface (provided we have the correct wifi
credentials set-up) when an ethernet cable is already offering connection to the board.

Without this change, in the case when ethernet is detected before wifi, connman would not
configure the wireless interface to get an IP. It would just configure and assign IP to wired
according to network set-up and then stop. This happens when ethernet has a higher preference
to wifi in PreferredTechnologies setting (or by default if PreferredTechnologies is not specified at all).

This chance will not prevent the wired interface to also get an IP address because
connman treats ethernet as a special case:
( http://git.kernel.org/cgit/network/connman/connman.git/tree/doc/service-api.txt#n40 )
it always connects the ethernet interface when a cable is plugged in.

This change will also determine the default route: if the wireless interface has internet
connectivity, then it will become the default route. If it does not have internet connectivity
and ethernet is plugged in, default route will be set to the ethernet interface in case this
one has internet connectivity.

For a description on how PreferredTechnologies works, please see:
https://01.org/connman/documentation

Signed-off-by: Florin Sarbu <florin@resin.io>