resin-init-flasher: Remove board specific bootloader code
We do not handle changing the bootloader config files from within the flasher
anymore. It is now the responsability of the layer to provide a bootloader configuration
file (for example uEnv.txt_internal for a BeagleBone Black) which would be simply copied
over from the flasher to internal media for future internal media boot-ups.

Signed-off-by: Florin Sarbu <florin@resin.io>