image_types_resin: Use FAT16 for resin-boot and resin-conf
mkfs.vfat is configured to autoselected the optimal FAT size based on cluster
number. This autoselection is only done for fat12 and fat16. The maximum fs
size for fat12 is around 16M which both resin-boot and resin-conf exceed. This
means that as long as we don't go below 16M, mkfs will always select fat16 as
optimal FAT size.

This PR adds changes for partition type at creation time to match the ones
autoselected later by mkfs.

Signed-off-by: Andrei Gherzan <andrei@resin.io>