oe-release: Make all values not matching [A-Za-z0-9]* be guarded by quotes
Variable assignment values must be enclosed in double or single quotes if
they include spaces, semicolons or other special characters outside of
A–Z, a–z, 0–9. Shell special characters ("$", quotes, backslash, backtick)
must be escaped with backslashes, following shell style. All strings should
be in UTF-8 format, and non-printable characters should not be used. It is
not supported to concatenate multiple individually quoted strings. Lines
beginning with "#" shall be ignored as comments.

https://www.freedesktop.org/software/systemd/man/os-release.html

Signed-off-by: Andrei Gherzan <andrei@resin.io>