image-resin.bbclass: Move here depmod'ing logic for compressed kernel modules.
We move the depmod-ing here from resin-image.bb so that it applies to all resin
images: resin-image and resin-image-flasher.

In addition to that, <do_rootfs[postfuncs] += "generate_compressed_kernel_module_deps">
operation creates the correct modules.dep in the rootfs however this file is not present
in the final sdcard image because the sdcard image is created at the same time this
postfuncs routine is called.

Therefore we add this generate_compressed_kernel_module_deps function to ROOTFS_POSTPROCESS_COMMAND
to make sure the build system calls it once the rootfs has been created and before sdcard image is
done, making sure it is persisted through to the sdcard image.

Signed-off-by: Florin Sarbu <florin@resin.io>