prepare-openvpn.service: Add dependency on resin-device-uuid.service
We moved the dependency on resin-device-uuid.service from openvpn-resin.service
as this is really the component that depends on the service of the uuid generation
script.

We also change the dependency type from "Wants" to "Requires" so that we fail when the
dependencies fail because we really depend on what they do/generate.

Lastly, remove obsolete "Before" line as our openvpn-resin.service explicitly waits
on prepare-openvpn.service

Signed-off-by: Florin Sarbu <florin@resin.io>