resin-image: Generate fingerprints just before image creation
If we generate the fingerprints as a post rootfs, other functions might
change the rootfs in different ways (for example zap passwd). In this
case the fingerprints will depend on the order of this function and if
it is not the last one might include bad hashes. Moving this generation
before image create we ensure that all the other postrootfs functions
were finalized.

Signed-off-by: Andrei Gherzan <andrei@resin.io>