networkmanager: add NetworkManager. disabled by default
NetworkManager is configured to use a resolvconf-type script, but for
that to work we must set 'dns=default'. We also want NetworkManager to
leave the docker* and tun* interfaces alone, so add it to the
'unmanaged-devices' list.

The --with-resolvconf configure option to NetworkManager enables some
code that's surrounded by #ifdefs. It appears untested, and buggy.
gio_gio.h.patch fixes the bug.

resolv.conf.sh is the minimal resolvconf replacement. Don't use the real
resolvconf, which is rather advanced, as it introduces unwanted complexity
when switching back and forth between NetworkManager and connman.

Signed-off-by: Michal Mazurek <michal@resin.io>