docker-disk: Tune options to the mkfs.btrfs command
Enable mixed mode (--mixed) to remove the isolation of data and metadata
block groups. The separate allocation of block groups leads to a
situation where the space is reserved for the other block group type, is
not available for allocation and can lead to ENOSPC state.

Set metadata profile to single (--metadata=single). The filesystem will
only store one copy of metadata.

Signed-off-by: Theodor Gherzan <theodor@resin.io>