host-update-resin.bbclass: Use a FHS at the root of the resinhup archive
In order for the archive to be a valid docker image, it has to have the
userspace unpacked and at the root of the archive. The boot files will be
stored also at the root directory, under a well known path that is unlikely to
have a collision with the actual contents of the root partition. Using
/resin-boot sounds like a safe option.

Signed-off-by: Andrei Gherzan <andrei@resin.io>