resin-supervisor.service: Move the docker container removal in ExecStopPost
ExecStopPost commands will be called after the service is stopped. This includes cases when
the commands configured in ExecStop= were used, where the service does not have any ExecStop= defined,
or where the service exited unexpectedly. This is the place to have the container removal instead of in
ExecStop because the commands specified in ExecStop= are only executed when the service started successfully
first.

We also tweak the ExecStop command in the sense of adding a 2 seconds timeout after which docker will send
the SIGKILL signal.

Signed-off-by: Florin Sarbu <florin@resin.io>