packagegroup-resin.bb: Adapt for new package dependencies
We will start to use aufs as docker storage driver with ext4 as
backing filesystem instead of btrfs.

For this to happen, resin-btrfs-balance recipe has been obsoleted and removed.
Also, we now use a new resin-filesystem-expand recipe for handling the ext4
filesystem expanding in the resin-data partition as well as we have renamed
resin-extend-expand to resin-data-expander for clarity.

Signed-off-by: Florin Sarbu <florin@resin.io>