networkmanager: add and configure NM 1.4.0
Use NetworkManager 1.4.0, meta-openembedded only ships with 1.0.12.
The recipe for 1.4.0 is based on the one from 1.0.12, and is as close to
it as possible. The modifications went into the corresponding .bbappend file.

NetworkManager is configured to use resolvconf, but for that to work we must
set 'dns=default'. We also want NetworkManager to leave the docker* and tun*
interfaces alone, so add it to the 'unmanaged-devices' list.

Signed-off-by: Michal Mazurek <michal@resin.io>