networkmanager_%.bbappend: Fix appending to PACKAGECONFIG
Using "PACKAGECONFIG +=" for networkmanager does not actually append but rather just
sets the PACKAGECONFIG variable to these values. This is because the PACKAGECONFIG
variable in the networkmanager recipe is set using "??=" (weak default assignment).

The operator "_append" differs from the operator "+=" in that it is deferred until
after parsing completes rather than being immediately applied. Also, the assignment
"??=" is made at the end of the parsing process (and only if the variable has not been
set in the meanwhile, hence the name weak default assignment) rather than immediately
thus "_append" is what we need in this case.

Signed-off-by: Florin Sarbu <florin@resin.io>