resin-logs: Reimplement persistent logging mechanism
This works as follows: script will check first for the config.json flag and set the journald.conf
accordingly.

The services depends on:
- boot mount point because it needs config.json
- var-log-journal because it needs this location writable in case of persistent journald
- a journald.conf.d bind mount so it can overwrite journald.conf

The prepare service does only a mkdir and it is needed because /var/log is a symlink to
volatile/log where /var/volatile is a tmpfs. So we need to make sure we create a needed
directory in the new tmpfs location. systemd tmpfiles was avoided because we want to
introduce as few dependencies as possible before we can start journal daemon.

As well this package adds a drop in file for journald service so it will start after it
is configured by this resin-logs.

Signed-off-by: Andrei Gherzan <andrei@resin.io>