docker: Use dnsmasq resolver
The host's /etc/resolv.conf has only one entry which is 127.0.0.2 which is a
local dnsmasq resolver. When docker starts a container it filters it out since
it can't be used from inside the network namespace of a container and because
there are no nameservers left it uses the default ones which are Google DNS.

We want the container to use the dnsmasq resolver as it can handle DNS much
better than the plain glibc resolver and we also want the containers to use the
DHCP provided one in addition to Google DNS.

One way to do that is to set the --dns <docker bridge IP> parameter when we
start the docker daemon so that every container has the docker bridge in its
/etc/resolv.conf. Then we have to forward any DNS packets going to the bridge
to 127.0.0.2 which can be done with iptables. Making dnsmasq itself listen on
the bridge interface means it won't be up before docker which is undesirable.

Signed-off-by: Andrei Gherzan <andrei@resin.io>