resin-init-flasher: Adapt for hardware needing the bootloader to be split into first stage bootloader and second stage bootloader.
Some hardware requires the use of a MLO (a.k.a. SPL - secondary program loader) that
is to be copied in static RAM and executed from there (static RAM is small in size)
and this first stage bootloader is responsible for initializing the regular RAM and
then copying the regular bootloader to this regular RAM and passing execution to it.

For this purpose we now allow a second set of variables called BOOTLOADER_FLASH_DEVICE_1,
BOOTLOADER_IMAGE_1, BOOTLOADER_BLOCK_SIZE_OFFSET_1 and BOOTLOADER_SKIP_OUTPUT_BLOCKS_1
to be used to accomodate this use case.

Signed-off-by: Florin Sarbu <florin@resin.io>