meta-resin-compat: Introduce new layer
meta-resin-common will target one of the latest poky versions available. In doing so,
it will inevitably hold code which will be incompatible with older poky versions. Hence
we introduce a new layer to our directory structure, meta-resin-compat, which will hold
all recipes/classes that are to be used by older poky versions.

Signed-off-by: Florin Sarbu <florin@resin.io>