meta-resin-compat os-release: Add to compat folder
os-release from meta-resin-common now contains yocto morty specific code so
for compatibility with older poky versions we add the older os-release recipe
to the compatibility layer.
We also BBMASK the meta-resin-common os-release bbappend so that when
meta-resin-compat will be added to the BBLAYERS then we are sure the
os-release bbappend from it is used instead of the one from meta-resin-common.

Signed-off-by: Florin Sarbu <florin@resin.io>