meta-resin-compat dnsmasq: Add recipe to compat layer
dnsmasq from meta-resin-common is now a bbappend for a newer dnsmasq
recipe which uses source code that does not need a certain patch applied.
However we need to be compatible with older meta-openembedded versions too
so we add the older dnsmasq bbappend and patch to the compatibility layer
and at the same time BBMASK the meta-resin-common dnsmasq to make sure users
of meta-resin-compat use the correct dnsmasq recipe.

Signed-off-by: Florin Sarbu <florin@resin.io>