networkmanager_%.bbappend: Enable gobject-introspection
Poky morty has support for GObject introspection data.
Without enabling introspection in NetworkManager, the build will fail with the error:

| make[4]: *** No rule to make target 'NM-1.0.typelib', needed by 'all-am'.  Stop.

This commit enables gobject-introspection inheritance to get rid of the build error and
as we do not use GObject introspection, we do not add the introspection data to the main
package, but rather add it to the libnmglib package.

Signed-off-by: Florin Sarbu <florin@resin.io>