meta-resin-compat dropbear: Add recipe to compatibility layer
The dropbear recipe in meta-resin-common now includes
a modified patch which makes it compile with poky morty
but make it fail to compile with older poky versions. So
in order to have the dropbear patch compile with older
poky versions we add the older dropbear code to the compat
folder and also BBMASK dropbear from meta-resin-common so
that meta-resin-compat users don't accidentally compile the
new incompatible code.

Signed-off-by: Florin Sarbu <florin@resin.io>