resin-supervisor: Add missing dependency
As the update-supervisor script changes the supervisor
configurations to our persistent location (through the bind mount), this changes need to be made
available to the supervisor, so we do not end up with two sources of truth,
i.e. the information found in the supervisor.conf after a supervisor update
(the one in our persistent location) will differ than the ones found in
the default location (read only location).

Signed-off-by: Theodor Gherzan <theodor@resin.io>