image_types_resin.bbclass: Conditionally use the resin image types from IMGDEPLOYDIR instead of DEPLOY_DIR_IMAGE
Because of oe-core commit OE-Core 6d969bac - "image: Deploy images to IMGDEPLOYDIR"
we need, for poky morty or newer, to reference the image types from a different directory
(IMGDEPLOYDIR instead of DEPLOY_DIR_IMAGE - in resin-image case this is a directory called
deploy-resin-image-image-complete from within the image's workdir).

Regardless of this change, each image type will be deployed also to DEPLOY_DIR_IMAGE
when an image is fully built (as part of the image's do_image_complete task).

Signed-off-by: Florin Sarbu <florin@resin.io>