run-resinhup: Implement cached docker image pulls
Based on a command line argument, have the ability to cache the docker images
and reuse them when possible. After the first pull the wrapper will save the
image as a tar archive (gzip) and next time it will load it avoiding pulling
over network.

This feature should be used only in demos, testing or development iterations.

Signed-off-by: Andrei Gherzan <andrei@resin.io>