meta-resin-morty linux-firmware_git.bbappend: Revert firmware update for rt2870.bin
With version 0.36 of rt2870.bin, connman 1.33 and wpa-supplicant 2.5,
the wireless chipsets using this firmware file are not always successful
in getting an IP address for the wireless interface. Getting an IP address
is randomly successful unless we explicitly do a "connmanctl disable wifi"
followed by a "connmanctl enable wifi".

In order to not mess around with connmanctl, we revert to the working
0.29 firmware version.

Signed-off-by: Florin Sarbu <florin@resin.io>