resin-filesystem-expand: Only expand if we have where to expand
The expander always needs to fsck before running, this operation being
time consuming we should avoid doing it when it is unnecessary. So we
shall determine before running the scan wheter we can expand or not based
on the remaning free block that can be found in the partition.
If we have at least one block free we shall expand.

Signed-off-by: Theodor Gherzan <theodor@resin.io>