resin-data-expander: Remove as replaced by an initramfs script
This script was moved to initramfs. This is because given that we now
mount the state partition in initramfs, resizing the extended partition
is dangerous after this mount happened (we mount machine-id in initramfs
which is on the resin-state logical partition in the extended primary
partition).

This functionality was implemented in resindataexpander initramfs
script.

Signed-off-by: Andrei Gherzan <andrei@resin.io>