kernel-resin.bbclass: Fix select flags for audit
We disablthe audit subsystem to avoid flooding dmesg.
Even if we try to explicitly disable the audit subsystem, there are
flags (i.e. apparmor && security && net) that togheter will still force
audit to be enabled.
We choose to disable the security flag as we do not use any
specific kernel security.

Signed-off-by: Theodor Gherzan <theodor@resin.io>