docker-resin-supervisor-disk: Skip the check that uses "docker images" for the supervisor image
The recipe checked for the presence of the supervisor image by using docker images and a regex. This was unstable,
so we remove it and the check will still happen with a "docker inspect" that is done to fetch the version. If that fails we'll
get a "Cannot fetch version" error that will be in this case indicative or a missing supervisor image.

Signed-off-by: Pablo Carranza Velez <pablo@resin.io>