systemd: Don't use the getty generator even in development images
systemd-serialgetty handles explicit getty spawns so even in development
this generator is useless. In the same way it enables us to have a
console argument different than where we spawn getty.

This will help when using serial0 aliases with raspberrypi boards. The
bootloader in that case replaces the from cmdline serial0 with the
actual device node and systemd will start getty in it. In uderspace if
we use serial0 aliases (symlinks from udev) and start getty in it we
will end up having the service on the same device twice.

Signed-off-by: Andrei Gherzan <andrei@resin.io>