openvpn: Switch to using a NetworkManager based service
Switching the resin-vpn system to be NetworkManager based allows
the networking to be centrally controlled and configured.  You
can bind the vpn to be brought up and down with specific interfaces
and connections by adding the following like to the [connection] section

secondaries=df32c2c2-2000-11e7-b51d-d050990ac729

This has been added to the resin-sample as an example.

Beyond that we are using a similar mechanism where the prepare-openvpn script
is parsing the config.json and then substituting necessary values into the
resin-vpn networkmanager connection script and moving it over to the rootfs.

Signed-off-by: Jon Nettleton <jon@solid-run.com>