resin-image: Use a standalone image to avoid IMAGE_ROOTFS_MAXSIZE hack
This image was based on core-image minimal which adds a 4M as extra rootfs
space. This led to a hack which was setting IMAGE_ROOTFS_MAXSIZE dynamically
based on IMAGE_ROOTFS_SIZE (+ 4M). We can avoid this by just not relying on
core-image-minimal and have resin-image as a full image recipe.

Signed-off-by: Andrei Gherzan <andrei@resin.io>