plymouth: resize to full screen instead of half screen
Previously the plymouth script resized the logo to 50% of screen.
Remove this behaviour, and resize to the full screen, continuing
keeping the image aspect ratio.

Also resize our logo to add transparent borders, so that the default
splash screen does not appear to change for users who don't replace
their splash.

Signed-off-by: Gergely Imreh <imrehg@gmail.com>