containers: implement mkfs for hostapps
This is a special filesystem creation tool that gets parameterized with a
docker image instead of the usual filesystem tree. The process is using the
native docker daemon produced by the build to load the image into the target
filesystem.

The tool uses the same hostapp-update tool used to update the target in order
to create the appropriate filesystem structure.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>