gnutls: Update to 3.5.9 in morty
The NetworkManager loads the gnutls library which calls on initialisation getrandom(). This call blocks until the entropy level is high enough in the system to initialise the random number pool, see #831. This delays the start of the NetworkManager, and in our case until the timeout of 1.5 minutes is reached. There is no need for gnutls to call getrandom() during initialisation because at this stage no en- or decryption operation is called. In gnutls 3.5.9 and newer, this issue is fixed. Unfortunately, this version isn't part of morty, yet. Therefore, if you update to NetworkManager 1.8, then you should also update the gnutls library.

Changelog-entry: Update gnutls to 3.5.9 in morty because of blocking NetworkManager
Change-type: patch
Signed-off-by: Robert Fritzsche <r.fritzsche@gridx.de>