mkfs.hostapp-ext4: Make sure PATH contains entries to container's paths
We use a yocto's PATH to pass it in the container for it to have access to the
sysroot paths. Since pyro, yocto has per recipe sysroots so a recipe will not
have access to a shared sysroot. In the same time, yocto's path excludes host
paths to avoid host contamination but pasing this path in the container will
make the container not able to find executables in the docker images based on
PATH. This patch makes sure that the PATH passed to the container contains
entries to host paths too.

Change-type: patch
Changelog-entry: Fix mkfs.hostapp-ext4 with pyro
Signed-off-by: Andrei Gherzan <andrei@resin.io>