docker-resin-supervisor-disk: allow supervisor start when no supervisor image present
The previous logic did not take into account what happens if there
is no resin-supervisor image nor container on the device, and
could end up in a state that the device cannot automatically recover
itself from a missing supervisor image. This happened for example
for resinHUP 1.x->2.x, when after the update it was necessary for
the device to start up without any images or containers.

Add logic to handle the missing image case by running the
update-resin-supervisor service when no image exists.

Also added comments so that the logic is documented.

Change-type: patch
Changelog-entry: fix start-resin-supervisor to start even when no supervisor image present
Signed-off-by: Gergely Imreh <imrehg@gmail.com>