image-resin.bbclass: Add sanity check for the boot partition
When we run host OS updates, we use atomic copy operations for updating the
boot partition. This requires at least free space as much as the largest file
in the partition. This check makes sure that the boot partition has enough free
space to do atomic copy operations.

Some of the work is based on:
http://git.yoctoproject.org/cgit.cgi/poky/tree/meta/classes/image-live.bbclass#n179

Change-type: patch
Changelog-entry: Check for enough space in boot partition to support atomic copy operations
Signed-off-by: Andrei Gherzan <andrei@resin.io>