hostapp-update-hooks: Add a boot files size check for 0-bootfiles hook
We want to "know" if the copy operations to boot partition will succeed before
starting to do it. This implementation adds a dry run which makes sure that at
no point in time the boot partition is left with less than a specific treshold
defined in the code.

Change-type: patch
Changelog-entry: Add boot partition space check for 0-bootfiles hostapp-update hook
Signed-off-by: Andrei Gherzan <andrei@resin.io>