resin-device-register: Delete the provisioning key after registering
This should prevent the VPN using the provisioning key to connect (which will not work).
As soon as the device is provisioned, the deviceApiKey will be used for the VPN to connect.
This is better than changing prepare-openvpn to not use the provisioning key at all, because
in resinhup scenarios the deviceApiKey may not yet work, but in resin-device-register, which
only runs in the flasher, we're guaranteeing that the deviceApiKey has been registered.

Change-Type: patch
Changelog-entry: Delete the provisioning key after registering, to fix VPN authentication in flasher images
Signed-off-by: Pablo Carranza Velez <pablo@resin.io>