hostapp-update: Run the update container as privileged
Because as part of a host OS update we need to update the bootloader also,
for some boards we need to be able to write the bootloader directly to a
specific /dev entry. Hence we need the update container be able to have access
to /dev of the host.

Change-type: patch
Changelog-entry: Run the update container as privileged
Signed-off-by: Florin Sarbu <florin@resin.io>