docker-resin-supervisor-disk: Don't manage supervisor with systemd
With the work of healthcheck balena might restart supervisor in case of an
issue detected. This would mean that systemd will have to not trace the balena
run process. This needs changes in update-supervisor too as it will need to
manage the supervisor with balena and not with the systemd service.

Change-type: patch
Signed-off-by: Andrei Gherzan <andrei@resin.io>