hostapp-update: make hostapp-update-hooks rollback on failure optional
When updating from a system that does not have hostapp-update-hooks or
not capable of rollback, have a flag that allows the script to skip
that step on failed hooks run, as it should be safer from the system's
point of view. Try to fix any issues manually instead of automatically
which in those circumstances has the potential to make things worse.

Connects-to: #939
Changelog-entry: Make hostapp-update-hooks rollback on failure optional
Change-type: patch
Signed-off-by: Gergely Imreh <imrehg@gmail.com>