networkmanager: Initilize resolvconf file with 8.8.8.8 nameserver
When running in the container, nm will ignore veth interfaces- which provide
the connectivity in such a case. This will make NM not create the resolvconf
file which breaks dns resolution completely because dnsmasq uses that file as
resolv file.

We fix this by making systemd tmpfiles service initialize the nm resolvconf
file with 8.8.8.8 nameserver.

Signed-off-by: Andrei Gherzan <andrei@resin.io>