plymouth: Mask broken services
These service are generated by the build system as broken services which throws
errors at boot and makes systemd ignore them. This is because at build time,
plymouth can't detect systemd ask password agent. This is a bug in meta-oe but
as we don't take advantage of these services in resinOS, we will mask them
altogether.

Signed-off-by: Andrei Gherzan <andrei@resin.io>