resin-filesystem-expand: Avoid this service in a docker container
Device units never activate in a container due to:
https://github.com/systemd/systemd/issues/6569. This makes this service always
fail when running resinOS in a container, and because this service is Before
mnt-data.service, this delays the boot process by 4m (until the device unit
fails). In order to fix this in the container we remove the dependency on the
device unit and to not afect the availability of the target label (resin-data)
when running on physical devices, we add dependency on udev settle.

The second problem is that we don't what this service to run at all because
when resinOS is ran as a container, the mnt-data is a volume. We avoid this
service in such a case using a service condition.

Change-type: minor
Changelog-entry: Add support for running resinOS in a docker container
Signed-off-by: Andrei Gherzan <andrei@resin.io>