healthdog: Fix RPATH on newer meta-rust
New meta-rust version don't set RUNPATH anymore and the current fix was
patching it only when the RUNPATH was set.

This patch is based on
https://github.com/meta-rust/meta-rust/commit/463622c0c6828e7fb970f46f9bf0dcf156345373
which is only available for recipes which use the rust-bin bbclass. Healthdog
uses cargo so the fix is not available for it. Furthermore, the fix is only
available on master (not on pyro, the branch we are working on).

Change-type: patch
Signed-off-by: Andrei Gherzan <andrei@resin.io>