resin-image-initramfs: Increase the MAXSIZE to 12M
By adding fat fsck we get above 8M limit of initramfs.

For example on rpi:
The initramfs grows from 8180K to 8236K and the final kernel image 8666232 to
8694552 (28K). Even so, we can only increase the maxsize by multiple of 4M due
to how yocto aligns fs to 4M for performance purposes so we need to increase
from 8 to 12M.

This only modifies the max size of the initramfs not of the rootfs which still
obeys the resin partition limit.

Signed-off-by: Andrei Gherzan <andrei@resin.io>