docker-disk: Various improvements
Currently we populate the docker disk partition by using a loop device and the
build outputs a fs file owned by root. We can do better:
1. stop using loop devices
2. have compile populate a directory which includes the content of the docker
partition
3. add a task which creates the filesystem out of the directory populated
above (after compile before install)
4. create the filesystem with fakeroot (chowning to root) so the final fs will
have content owned by root

This work was part of the work done for #114 which was aiming to run the
docker container which creates the docker partition unprivileged. Currently
that is not possible due to the fact that we run docker in docker which
requires elevated permissions.

This patch fixes as well a racing issue where multiple builds might have
collied when creating the looper imager (was renamed as well to be clearer what
it is). Now we create images that are unique per build and we cleanup
afterwards the image.

Change-type: patch
Change-log-entry: Various build improvements for docker-disk recipe
Signed-off-by: Andrei Gherzan <andrei@resin.io>