kernel-modules-headers: Compile depends on do_patch
When compiling this package, the tool uses the kernel source directory
STAGING_KERNEL_DIR along with the deploy dir (which includes artifacts from
building the kernel). We already depend on do_deploy for kernel but when building
from sstate, the STAGING_KERNEL_DIR is not populated. Make sure that is
populated when compiling kernel-modules-headers by adding an explicit
dependency to kernel:do_patch.

Change-type: patch
Changelog-entry: Fix kernel-modules-headers when building from sstate
Signed-off-by: Andrei Gherzan <andrei@resin.io>