resin-mount: Rename partition services and fix service start/stop
After merging the in container resinOS PR
(https://github.com/resin-os/meta-resin/pull/951) we broke mounting the state
partition on boards. This is due to
https://github.com/resin-os/meta-resin/issues/975 . This change fixes the issue
above by adding a wrapper which checks if a certain label is already mounted.
It handles as well umounts so we can stop the partition services as well. These
services were updated to use this mount wrapper (reisn-partition-mounter).

The second change this introduces is the renaming from mnt-(*).service to
resin-(*).service. The old naming convention was inherited from the way systemd
works with mount units and now, given that we moved to services, we can use
clearer service names for these partition mount services.

Fixes #975
Fixes #213

Change-type: patch
Changelog-entry: Rename partition mount services to <label>.service and handle stopping services
Signed-off-by: Andrei Gherzan <andrei@resin.io>