modemmanager_1.7.990.bb: Add dependency to libxslt-native
When compiling with Poky Rocko we need to explictily add a build-time
dependency to libxslt-native.

This fixes the following error:

| Making all in include
| make[2]: Entering directory 'build/tmp/work/core2-64-poky-linux/modemmanager/1.7.990-r0/build/include'
| xsltproc --xinclude --nonet ../../ModemManager-1.7.990/build-aux/header-generator.xsl ../../ModemManager-1.7.990/introspection/all.xml > ModemManager-names.h
| /bin/bash: xsltproc: command not found
| Makefile:648: recipe for target 'ModemManager-names.h' failed
| make[2]: *** [ModemManager-names.h] Error 127

This is introduced by
http://cgit.openembedded.org/openembedded-core/commit/?id=f7f87b8840f5997f65ddf643f26dde0ed5f6c308

ModemManager inherits vala and because we do not use the "api-documentation" DISTRO_FEATURE the conditional
dependency on libxslt-native introduced by the above commit is not triggered so we need instead to depend
explicitly ourselves on libxslt-native.

Change-type: patch
Changelog-entry: Make ModemManager depend on libxslt-native at build-time for Poky Rocko
Signed-off-by: Florin Sarbu <florin@resin.io>