dnsmasq: Use update-alternatives for deploying resolv.conf
In our build, both dnsmasq and systemd want to provide resolv.conf.
Because dnsmasq is an additional resolv.conf provider, let's use
update-alternatives for this while also defining a priority of 60,
above the 50 from systemd, which means resolv.conf on the target will
be a symlink to the resolv-conf.dnsmasq provided here.

Without using this mechanism, on Poky Rocko, the build system will error out
because systemd will be registered as a postscript to be run on target for
doing this symlink and this shall fail due to our read-only rootfs:

ERROR: resin-image-flasher-1.0-r0 do_rootfs: The following packages could not be configured offline and rootfs is read-only: ['systemd']
ERROR: resin-image-flasher-1.0-r0 do_rootfs: Function failed: do_rootfs

Change-type: patch
Changelog-entry: Use update-alternatives for deploying resolv.conf through dnsmasq
Signed-off-by: Florin Sarbu <florin@resin.io>